requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
altair==5.4.1
asttokens==2.4.1
async-timeout==5.0.1
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
dateparser==1.2.0
DateTime==5.5
debugpy==1.8.8
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
fonttools==4.55.0
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.1.0
narwhals==1.14.0
nest-asyncio==1.6.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pycryptodome==3.21.0
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
python-binance==1.0.22
python-dateutil==2.9.0.post0
pytz==2024.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
streamlit==1.40.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.3
watchdog==6.0.0
wcwidth==0.2.13
websockets==14.1
yarl==1.17.1
zope.interface==7.1.1
